aaaaaaa

+++ backed/database/model/verificationTokens.go
@@ -6,7 +6,7 @@ import (
 
 type VerificationToken struct {
 	Token     string `gorm:"uniqueIndex;not null"`
-	UserID    uint   `gorm:"not null"`
+	UserID    int64  `gorm:"not null"`
 	User      User   `gorm:"constraint:OnDelete:CASCADE;"`
 	ExpiresAt time.Time
 }
